Split conversion-related code out of host into a new file.
